ToMayaMeshConverter : No longer set normals Currently, when converting from a Cortex object to a Maya Mesh, we set the normals anytime the normals are stored in the scc. Setting the normals explicitly implies these are static values and so Maya locks them rather than computing them on the fly.
Locked normales are undesirable in many parts of the Pipeline,
especially for deforming geometry as not recomputing them leads
to unexpected results, e.g. wrong tangents or incorrect shader
computations.

This change removes the normal transfer from the scc to the
Maya geometry and instead we rely on Maya to compute them.