VDBObjectTest : Tweak for OpenVDB 10.1
The grid value class bindings are intended to provide everything both as an item (`["value"]`) and as an attribute (`.value`), but in OpenVDB 10.1 the item bindings seem broken, so we must use the attribute instead.

Also added a missing import, and used `next()` to get the first element from `citerAllValues()`, knockin 2s off the runtime of the test.