IECoreBinding : Use `Py_ssize_t`
MSVC has `SSIZE_T`, but not `ssize_t`, which is not part of the standard
it seems. Python's `pyport.h` file does a `typedef` for it, but
something must have changed between Python `3.7` and `3.10` that causes
that to not be included from what we're using. It does seem that the
official recommendation is to use `Py_ssize_t` now.