IECore module : Add DLL directory for DLL dependencies
This fixes `scons testCorePython` and provides a mechanism for folks to fix
imports in their own environments.

At first I tried to just hardcode the directory to match the directory
layout used by Gaffer, using a path relative to `__file__`. But that didn't
work, because when we run `scons testCorePython`, we haven't installed the IECore
module yet (we test it in place in the local dir). The `IECORE_DLL_DIRECTORIES`
environment variable was the only other option I could see, and I suppose it
at least gives folks a hand if they have their own installation layouts.