USD ShaderAlgo : Fix `treatAs*` and `arnold:*` light parameters
The main motivation for this is to allow Arnold-specific parameters written out of Gaffer to render as expected via the `arnold-usd` render delegate. In Gaffer we represent these as standard light parameters named as `arnold:{name}`. But `arnold-usd` made the unfortunate decision not to represent them as standard light inputs in USD, but to store them as primitive variables instead. With a bit of luck, `arnold-usd` will align with the USD convention in future and we can ditch this code.