ShaderStateComponent : Handle ToGLTextureConverter exceptions
These were thrown if a GLSL shader had bad values for a texture parameter - the case that motivated this was a custom setup in production that would pass a CompoundData with no channels and an empty data window.

We can't let the exception propagate because `addParametersToShaderSetup()` is called during drawing and the exception would prevent the rest of the scene drawing and/or lead to a corrupted GL state.