USD ShaderAlgo : Emit warning for unsupported parameter types
In conjunction with the previous commit, this means that we now gracefully print a warning message for any parameter types we can't write to USD. Prior to this we were throwing a not very helpful exception and not writing anything at all. This came up in production with a GLSL shader with an ImagePlug input in Gaffer, where we pack an entire image into a CompoundData value.