CI : Update Node12 actions
This matches the action versions used on Gaffer CI and silences the "The following actions uses node12 which is deprecated and will be forced to run on node16" warnings.

There are even newer versions of these actions, such as checkout and upload-artifact v4, but they require Node20 which we're unable to use on our CentOS 7 build container.