OpenImageIOAlgo : Add missing array handlers to `data()`
And consolidate the existing array handling to use the same templated function as everything else. This new version is preferable because it doesn't initialise the data twice (previously the `resize()` was doing a redundant initialisation prior to the `copy()`).

This fixes the loading of ICCProfile metadata, which is the underlying problem reported in https://groups.google.com/g/gaffer-dev/c/qUSJ-r0XuYI/m/y4OMtrBUAAAJ.