IECoreUSD : Use `/` for resolved paths.
On Windows, the result of USD's `GetResolvedPath()` method returns the
resolved path using `\` between directories. This can cause problems
when doing string substition on a resolved path since the `\` will be
interpreted as the start of an escape character. Using `/` also
conforms to Cortex's standard of always using `/` for paths because they
work on all operating systems.