USDScene : Fix round-tripping of colons in set names
And make some notes on future changes we'll want to make to our name mangling. In theory we could implement round-tripping _now_ either by using metadata to store the "true name" or by implementing the first USD proposal ourselves. But I don't think round-tripping of arbitrary names is pressing enough to warrant the extra effort and ugliness if USD will provide it for us in the medium term.

Note : I removed the first call to `validName()` in `writeSetInternal()` because in that case we're going to make the name valid in the recursive call we're passing the name to.