USD PrimitiveAlgo : Fix crash loading skinned facevarying normals
The call to `skinningQuery.GetJointMapper()` can not only reorder transforms but can also omit them. This occurs when SkelBindingAPI uses a `skel:joints` attribute to bind a mesh to a subset of joints, allowing the weights for other joints to be omitted. We were assuming that the mapped transform array had the same length as the input array, and were running off the end of the array.