ShaderNetworkAlgo : Allow customisation of component adaptors
The main motivation here is that the MaterialX shaders we were using as adaptors don't really make sense - they're not actually available as OSL shaders any more, so they have zero chance of working in other DCCs when we export shading to USD from Gaffer. Instead we can now register renderer-specific adaptors that are available everywhere the renderer is.

This also somewhat changes our approach to removing adaptors. Instead of just tagging them as "an adaptor" and figuring the rest out based on the shader name, we now register all the necessary information as blind data. This will allow us to reliably remove an adaptor even if adaptor registrations change in the future. It should also be useful at a certain studio which rewrites the `info:id` attribute on export to USD, in such a way that it broke our old adaptor removal code.