IECoreUSDModule : Make compatible with USD 25.05
In which the USD Python bindings are now performed using a copy of `boost::python`, housed in the `pxr_boost::python` namespace. We bind a few functions of our own that deal with `pxr` types, so we now need to teach our `boost::python` how to pass them to `pxr_boost::python`.