USD : Add loading of Arnold procedural prims
These are converted to ExternalProcedural objects in Cortex. I haven't implemented writing yet, mainly because we don't actually have a way of telling which renderer an ExternalProcedural is intended for at present. That's something we should probably sort out at the same time as sorting out the `set/getFileName()` confusion.