USDScene : Fix loading of instanced skinning with unique animation
The `skel:animationSource` can be inherited from ancestor prims, allowing instanced skeletons to receive unique animation. We need to account for that when generating the object hash, otherwise objects with distinct animation will falsely share a hash.

In the comment I've mentioned that ideally ObjectAlgo would be in charge of this hashing, but I'm not 100% sure that is the case. I have a prototype which does more accurate hashing using SdfPrimSpecs, and in that USDScene tracks an inherited hash to account for value clips. It may be that it would be better to track an inherited animation hash as well, which would be hard to delegate out to ObjectAlgo.