IECoreGL : Remove tests that depend on GLRenderer
We're about to remove the Renderer class and associated components, and these tests use them. This is the only part of this whole remove-the-renderer operation I feel bad about. There is some decent test coverage here, some of which is still relevent to the _new_ GLRenderer in Gaffer, which doesn't have as good coverage. On the bright side, it wasn't being run on CI anyway, and there's a good chance that any significant future overhaul of hardware rendering in Gaffer will mean a move away from OpenGL.