SceneCacheFileFormat: Add support for collection stored on the internalRoot.
USD doesnt support writing collections on the absolute root prim (`/`) but IECoreScene
does.
We have some tags we want to be able to write out of Houdini that should
be stored on root loacation (`/`), therefore we leverage the internal root to author the
collections and map that to IECoreScene root loaction.

Note that there is still a limitation for read tags from the root
location and converting them to collections stored on the internal root
but the current priority is to support writing out.