DisplayDriverServer : Support IPv4-only environments
In c357ea106daea2ee5aad6bb4f871ddb87e80991e, we switched from making an IPv4 connection, to an IPv6 one. This fixed some nasty lag on Windows, but it turns out that it also broke IPv4-only environments completely. So now we fall back to opening a v4 connection if the v6 connection fails.