SConstruct : Added INCLUDE_PATHS option and skip empty include paths
INCLUDE_PATHS is a list argument that can be used to add an arbitrary
list of include paths to be added, with full control over their order
from the options file.

We are also avoiding adding any include paths that are empty, when using
the old arguments.

That is not strictly necessary (we could be setting it to an invalid
path, for example) but does result in a much cleaner compile command.