CoordinateSystem: Backwards compatibility
CoordinateSystem used to derive from the deprecated and removed
StateRenderable holding additional transform information for the
renderer. A better approach that is also used in Gaffer, the transform
information is better stored as separate transform information.
Cortex 10.6 also removed all DCC integrations, i.e. IECoreMaya, which
stored, for example Locator transformation on the CoordinateSystem. We
provide a way to still retrieve this information from old SceneCaches,
although not fully transparent and backwards compatible.  Clients of
CoordinateSystem will need to be adjusted if needed.

The `M44fData` is stored as `LegacyTransform` in the `BlindData` of
the CoordinateSystem